rename method files
