+++ pyspw/methods/__init__.py
@@ -2,8 +2,8 @@ from .create_payment import CreatePayment
 from .create_transaction import CreateTransaction
 from .get_user import GetUser
 from .get_user_cards import GetUserCards
-from .me import Me
-from .selfcard_info import GetSelfCardInfo
+from .get_me import GetMe
+from .get_selfcard_info import GetSelfCardInfo
 from .set_transaction_webhook import SetTransactionWebhook
 
 
@@ -12,7 +12,7 @@ __all__ = [
     'CreateTransaction',
     'GetUser',
     'GetUserCards',
-    'Me',
+    'GetMe',
     'GetSelfCardInfo',
     'SetTransactionWebhook'
 ]
