+++ pyspw/methods/get_selfcard_info.py
